Add initial support for Simulator
This updates the code to compile against the latest closed beta for Simulator. None of the meshes we previously modified are readable, so only the rails are being regauged. All mesh names have been updated, so once that's fixed or we get the green-light to redistribute them like before, it should be good to go. Some exploration will have to be done to find any new meshes that need to be regauged.