Add Blender scripts for getting verticies
These scripts are from June 2023 and are what I used to get all the
verticies stored in the verticies.json file. I figured it's better
to have them checked in for anyone to use rather than continually
searching for them in Discord :D.

The copy script has a weird quirk of needing to switch back to object
mode before you run it or it'll copy nothing/stale verts.